+++ docs/tools-standard.qmd
@@ -410,7 +410,7 @@ inspect eval computer.py --model openai/computer-use-preview
 The computer tool supports the following options:
 
 | Option | Description |
-|----------------------|--------------------------------------------------|
+|-----------------------|-------------------------------------------------|
 | `max_screenshots` | The maximum number of screenshots to play back to the model as input. Defaults to 1 (set to `None` to have no limit). |
 | `timeout` | Timeout in seconds for computer tool actions. Defaults to 180 (set to `None` for no timeout). |
