+++ src/inspect_ai/tool/_sandbox_tools_utils/Dockerfile.pyinstaller
