+++ src/inspect_ai/tool/_sandbox_tools_utils/_build_bundled_executable.py
@@ -18,7 +18,7 @@ from pathlib import Path
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from inspect_ai.tool.sandbox_tools_utils._build_config import (
+    from inspect_ai.tool._sandbox_tools_utils._build_config import (
         SandboxToolsBuildConfig,
         filename_to_config,
     )
